--- src/components/organisms/Home.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Home() {
-  return (
-    <h1>Landing Page</h1>
-  )
-}
-
-export default Home
