+++ src/components/templates/AnimationPage.tsx
