+++ src/components/templates/SeriesPage.tsx
